Move the parser into a subdirectory.
